Move Contact to dedicated component
